Add files via upload
This was my attempt at making a program to transcribe audio files to written documents. A website called transcribeme offered money for transcripts made, so I thought I would try to write a program to do this as best as possible. It uses the Google api to translate audio to text. One problem that I ran into with using the Google api was that it would only accept small audio files. As a work around I created a program that would take an audio file of any length (.wav file) and split it into as many ten second segments that were needed. It would then send these segments to the Google api to be translated. The result was added to a list and saved. After all of the newly created files were transcribed, they were deleted. The output was saved to a text document called "Output".

What I learned from this project was that the Google api is not as accurate as I thought it was, often mixing up words if the file is not completely clear. It breaks the workload of any transcription down significantly as it gets a lot of  words correct, but human corrections are still needed afterwards. Overall it was a fun project but not the solution that I had hoped it would be.