Add files via upload
This was a simple key logging program that I made for fun.  Although the program can be adapted for malicious use, it was not intended for this and there are surely much better options online. It uses a couple of imported modules that act as event listeners, these event listeners wrote their output to a .txt file along with a timestamp for the user to later view. The program can also be terminated by pressing "insert" on the keyboard. I chose this key as it is on every keyboard size and is rarely used by the average person. This is best used with "pyinstaller" to turn the .py file into an exe which can then be used on any windows machine with ease without the need of a python installation.

During this project I learned how to have my programs run in the background of a machine without the average person knowing. (Even in .exe form it can still be found in task manager quite easily due to the name of the file and icon, so change that for best results.)